Refactor CardManager (#854)
* Add getIfShowOnStart() method to Card

If the Card should be displayed in the overview screen, this method returns the Card. Otherwise, it returns null.

* Add getCards() method to ProvidesCard classes

This method returns all Cards that should be shown in the overview screen. It’s not implemented yet in CafeteriaManager, as there are more changes to be made.

* Add synchronous query for cafeteria in CafeteriaViewModel

To achieve this, I refactored CafeteriaManager.Cafeteria into CafeteriaWithMenus.

* Improve code in CafeteriaWithMenus

* Add CardsRepository class which provides the list of Cards

* Add MainActivityViewModel that holds all UI data of MainActivity

* Update CardAdapter and refresh behavior in MainActivity

The CardAdapter now no longer gets its items from CardManager, but is update with items through adapter.updateItems(List<Card>).

* Add DiffUtil callback for updating the cards list view

* Fix problem in CafeteriaViewModel, remove unused code in CardManager

* Remove manual refresh CardManager.update() calls

* Add comment about future refactoring task