Fixed the crash in ShowTicketActivity after merge (#931)
* fix cut off image placeholder in drawer head (#876)

* fix cut off image placeholder in drawer head

* Change text, so it fits into one line

* Fix bug that occasionally caused transport search results to be empty (#882)

It occurred because the API occasionally does not return the "quality" property of a search result, resulting in a JSONException in the app.

* Fix crash in TransportDetails (#881)

The crash occurs when a subway departure does not contain the line
number. This seems to occur randomly at some transport stations, e.g.
Marienplatz.

* disable crashlytics for debug and -dev builds (#867)

* disable crashlytics for debug and -dev builds
this fixes #824

* Simplify disabling crashlytics.

* Add ConnectivityInterceptor to throw an exception if device is offline

The Interceptor throws a NoNetworkConnectionException, which is a subclass of IOException. You can react to it in Retrofit's onFailure callback.

* Add placeholder for news source icons

Before, the Facebook icon was set as the standard news source icon and was replaced when the actual icon was loaded. Now, we display a gray placeholder icon until the actual icon is loaded.

* Update design of CreateEventActivity (#886)

* Update design of CreateEventActivity

- Update paddings
- Update design of date and time selectors

* Update design of CalendarDetailsFragment

- Update paddings
- Update buttons

* Unify date formatting in CreateEventActivity and CalendarDetailsFragment

* Fix keyboard style for event title text field

* Make interface improvements

- Add click indicator for date and time text views
- Add hints to text fields

* Improve code formatting

* Make various user experience improvements

- Add close dialog to prevent accidental data loss
- Open keyboard automatically when creating new event
- Remove empty navigation drawer
- Disable save button if event title is empty

* Add German translation for close dialog text

* Hide keyboard when the user exits CreateEventActivity

* Add animations to GradesActivity (#837)

* add crossfade between chart types

* some code refactoring

* hide chart with animation + clearer code

* animate arrow + fix average grade

* different arrow animation

* fix rotation bug

* put paths in extra res values file

* fixes issue with refresh animation and overlapping of fading views

* Fix bug where TUFilm movies were downloaded but not handled successfully (#891)

When a new user tried to download TUFilm movies, they would receive a huge list of movies, which Observable could not handle. We're now using Flowable, which includes backpressure support and can handle a large number of emitted items.

* Fix DB index error on first launch (#880)

* Fix DB index error on first launch

* Ignore failing database tests

* Prevent crash caused by WifiManager (#883)

* Prevent crash caused by WifiManager

* remove NetworkChangeReceiver as suggested by @kordianbruck

* remove WifiScanHandler

* Check which ids/keys/tokens have been uploaded and upload missing ones (#902)

* rename gcm to fcm (firebase cloud messaging)

* model for checking which ids/keys/tokens have been uploaded to the server

* Add dedicated card layouts (#884)

* Add dedicated layout for tuition fees card

The tuition fees card previously used the general card_item.xml. This
apparently resulted in a crash when a dynamically added TextView was not
found.

To fix this, the tuition fees card now provides a dedicated layout
card_tuition_fees.xml with all its required views.

* Add dedicated layouts and ViewHolders for Cards that used card_item.xml

This is a first step towards making the existing way of populating ViewHolders easier and less error-prone. No longer uses card_item.xml, but dedicated layouts for each card type.

Also, add subclasses of CardViewHolder which are responsible for binding data to a card’s view.

* Add didBind flag to prevent content from being added multiple times

The content of MVVCard, ChatMessagesCard and CafeteriaMenuCard are added dynamically. To prevent the content from being added every time the ViewHolder becomes visible, I added didBind as a boolean variable.

* Switch to CafeteriaWithMenus in CafeteriaMenuCard

* Remove commented code

* Remove MVV plans functionality (#921)

* Update README.md

* Update README.md

* Update README.md

* fixed crash in ShowTicketActivity by changing the dummy values for now