Rewrite notification scheduling
Notifications about tuition fees, cafeteria menus and transit departures require computation directly before the notification (fetch tuition fee status, get menus of currently preferred cafeteria, fetch departures for currently nearby station). For that reason, we no longer schedule concrete notifications, but set an alarm for the desired time of the notification. Once that alarm goes off, we can use the corresponding NotificationProvider to fetch the actual content of the notification and display it.