Restore background calendar sync
In a previous refactor, I accidentally removed the background syncing
for calendar events. I've restored this functionality now.

This commit also catches SQLiteExceptions that occur during the
syncing of calendar events on Huawei devices. They are particularly
annoying to the user, as they often occur while the TUM Campus App is in
the background. After some research, there seems to be no obvious fix
for this problem.