Fix NullPointerException when clicking an event (#1023)
* Fix NullPointerException when clicking an event

When clicking an event in the week view, the corresponding CalendarItem is retrieved from the database. If the query returns null, the app crashes (happened very rarely). I’ve added a null check to prevent this.

Additionally, instead of using the event's start and end time to query for the CalendarItem, the proper event ID is now used.

* Remove experiment code