Unify cafeteria and favorite dish notifications (#1057)
* Add getIfShowOnStart() method to Card

If the Card should be displayed in the overview screen, this method returns the Card. Otherwise, it returns null.

* Add getCards() method to ProvidesCard classes

This method returns all Cards that should be shown in the overview screen. It’s not implemented yet in CafeteriaManager, as there are more changes to be made.

* Add synchronous query for cafeteria in CafeteriaViewModel

To achieve this, I refactored CafeteriaManager.Cafeteria into CafeteriaWithMenus.

* Improve code in CafeteriaWithMenus

* Add CardsRepository class which provides the list of Cards

* Add MainActivityViewModel that holds all UI data of MainActivity

* Update CardAdapter and refresh behavior in MainActivity

The CardAdapter now no longer gets its items from CardManager, but is update with items through adapter.updateItems(List<Card>).

* Add DiffUtil callback for updating the cards list view

* Fix problem in CafeteriaViewModel, remove unused code in CardManager

* Remove manual refresh CardManager.update() calls

* Add comment about future refactoring task

* Add AppNotification and NotificationsProvider taxonomy

* Add ProvidesNotifications interface to controller classes

* Adapt AppNotification classes and add notification scheduler, presenter, receiver

* Remove notification functionality from NotificationAwareCard and subclasses

* Update content of tuition fees notification

* Add improvements to display of notifications

Present news notifications in an expandable format.

* Improve display of cafeteria menus in notifications

* Add method to allow NotificationProviders to provide their own builder

* Improve display of cafeteria notifications

* Add more intelligent scheduling of cafeteria and tuition notifications

* Make API improvements in notifications package

- Changed the APIs of classes in the notifications package to allow for easier presentation and scheduling of notifications.
- Introduced a AppNotificationsManager that returns the list of classes the implement ProvidesNotifications
- Add documentation for classes in the notifications package

* Deprecate NotificationAwareCard and make subclasses inherit from Card

* Unify date formatting in tuition fees notification

* Add hasEnabledNotifications method in ProvidesNotifications

This method returns whether or not the user has disabled notifications for a particular notifications provider.

* Add various improvement to notifications package, remove unused code

* Remove deprecated NotificationAwareCard

* Make minor code formatting changes

* move setupNotificationChannels() to App
Notifications could get swallowed, if StartupActivity isn't run, which
would manifest in a bunch of Notifications being generated at startup

* refactor notifications to be a top level component
and also move specific notifications into components

* remove calendar notifications after the calendar entry finished

* set cafeteria notification timeout

* Added color (accent) attribute to all notification creation (#939)

* Fix merge problems

* Make minor code improvements

* Fix eduroam channel configuration

* Remove unused notification backgrounds

* Remove unused notification backgrounds

* Make minor code improvements

* Remove MVV notification option from settings

* Address code review feedback

* Rename regex patterns in CafeteriaMenu

* Fix merge error

* Make minor code improvement

* Rewrite notification scheduling

Notifications about tuition fees, cafeteria menus and transit departures require computation directly before the notification (fetch tuition fee status, get menus of currently preferred cafeteria, fetch departures for currently nearby station). For that reason, we no longer schedule concrete notifications, but set an alarm for the desired time of the notification. Once that alarm goes off, we can use the corresponding NotificationProvider to fetch the actual content of the notification and display it.

* Fix outstanding TODOs and remove commented code

* Remove commented code

* Unify cafeteria and favorite dish notifications

Now, the user can change the notification time of the “normal” cafeteria notifications. There are no longer separate notifications for favorite dishes. Instead, if a favorite dish is served on a particular date, the cafeteria notification will highlight it in the notification text.

* Remove unused method in CafeteriaMenuManager

* Fix merge problem

* Improve variable naming