Improve display of upcoming events in MainActivity
Instead of heading to CalendarActivity and displaying the CalendarDetailsFragment there, we display it directly in MainActivity. This way, the user can go back to browsing their feed more easily. The fragment now also contains a "Show in calendar" button, so that the user can open the CalendarActivity and perform actions on the event (such as edit and delete).